Created Branch for UTF8hotfix
The csv data had issues, that after filtering for date and saving
as utf8 format, there was a prefix to the first row.
This broke the writer and I patched it uglyly as I do, with
changing the key for the dicts.
This is only necessary in my case and hence i push this to a new branch.